Change implementation of applyTransition
Before we tried to find a suitable transition by looking
at the set of actions that end in a target state and select
one that starts in the current state. However, some transitions
have a special source state 'all' which means they can be applied
always. This caused a failure as the comparison between the source
state and the current state would fail.

There are many other reasons why an action could fail, therefore
it seems better to simply take the list of possible actions that
end in the target state and ask the SDK to validate them against
the current state of the issue. If the validation succeeds we perform
the transition.